### 2023-10-03 #5141 @arfon
@lohedges @symmy596 @hmacdope @lscalfi – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is[MD-SAPT: Python Based Toolkit for Running Symmetry Adapted Perturbation Theory Calculations on Molecular Dynamics Trajectories](https://github.com/openjournals/joss-reviews/issues/5141) – https://github.com/calpolyccg/MDSAPT

Many thanks
Arfon
