### 2023-10-03 #5141 @hmacdope
@arfon I will have to excuse myself on the basis of a COI, as I am an MDAnalysis core-developer (creators of the MDAKit ecosystem, which this package is a member of :smile: ), and have worked with Alia (the first author) in an open source capacity as a member of our dev community a few times. I believe this falls under the COI policy but you can correct me if I am wrong. :+1:
