### 2023-10-09 #5931 @editorialbot
[REVIEW]: MD-SAPT: Python Based Toolkit for Running Symmetry Adapted Perturbation Theory Calculations on Molecular Dynamics Trajectories
**Submitting author:** @armcdona (<a href="http://orcid.org/0000-0002-4381-1239">Ashley Ringer McDonald</a>)
**Repository:** https://github.com/calpolyccg/MDSAPT
**Branch with paper.md** (empty if default branch): JOSS_Pub
**Version:** v2.0
