### 2023-10-17 calpolyccg/MDSAPT#43 @CarvFS
Problem with installation
Hello! First of all: thank you very much for developing this tool!

I am here through [openjournals/issues#5931](https://github.com/openjournals/joss-reviews/issues/5931). I am installing your packaged and I am receiving an error message while running `mdsapt generate test`

```
Traceback (most recent call last):
  File "/home/felipe/miniconda3/envs/mdsapt/bin/mdsapt", line 5, in <module>
    from mdsapt.cli import cli
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/mdsapt/__init__.py", line 8, in <module>
    from .config import Config, load_from_yaml_file
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/mdsapt/config.py", line 132, in <module>
    class RangeFrameSelection(BaseModel):
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/mdsapt/config.py", line 145, in RangeFrameSelection
    @root_validator()
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/pydantic/deprecated/class_validators.py", line 228, in root_validator
    raise PydanticUserError(
pydantic.errors.PydanticUserError: If you use `@root_validator` with pre=False (the default) you MUST specify `skip_on_failure=True`. Note that `@root_validator` is deprecated and should be replaced with `@model_validator`.
```

My package versions are:

```
mdanalysis                2.6.1
mdsapt                    2.0.0+1.gffc3bd0
pyyaml                    6.0
pydantic                  2.4.0
```

Should I have other specific versions installed?

Also, I have installed by cloning your repository and then running `pip install ./MDSAPT`.
