### 2023-10-19 calpolyccg/MDSAPT#43 @ALescoulie
@CarvFS I looked into it and it seems like we need to modify our config validation due to pydantic deprecating `root_validator`. @ifd3f wrote most of config, and is going to get started on it soon.
