### 2023-11-02 #5931 @CarvFS
# Review: observations on the paper and functionality

First of all: thank you very much for developing this toolkit! It is important to make the non-convalent bond analysis easier for everyonw working with molecular dynamics. Here I will post some comments on the paper and toolkit. Any new observations I will post as a new comment.

## About installation:

I think the authors should update the documentation to avoid users running into issues while trying to install their toolkit. A few things I have observed:

- Python version 3.7 or older gives the error: "cannot import name 'Literal' from 'typing'";

- Trying to install as `conda install -c psi4/label/dev -c conda-forge mdsapt` is not working, it runs as
```
Retrieving notices: ...working... done
Collecting package metadata (current_repodata.json): done
Solving environment: unsuccessful initial attempt using frozen solve. Retrying with flexible solve.
Solving environment: unsuccessful attempt using repodata from current_repodata.json, retrying with next repodata source.
Solving environment: / Killed
```

- Trying to install by cloning the repository and running `pip install ./MDSAPT` also did not worked:
  - I had to install all necessary tools (e.g pydantic and MDAnalysis) manually. I bilieve those should be installed automatically along with MD-SAPT, since not having them will prevent the code to run successfully;
  - Them, I have got an error due to pydantic deprecating `root_validator`. I have openned an issue about this error.

Maybe using some specific version of pydantic would avoid my last error. However, it should be listed in the documentation which versions the user must have.

## About important links

- Link to readthedocs page at the paper is not redirecting to the correct page;

- The binder demo are returning some error messages. I liked very much that you have made it, so new users could just run it first without need to install anything. It would be very nice to keep it updated.

## About the paper and toolkit

- When it says "including" in the paper it gives the ideia that there are other tasks not described. Is this the case? Maybe describe the all workflow tasks as bullet points in the paper and/or add it to the documentation;

- It is said that a number of python based analysis tools exist for MD data. However, it is not clear if any of those also performs the same tasks implemented in MD-SAPT. For example: psi4 is used to run the quantum calculations, MDAnalysis is used to handle the MD simulation data. However, is not said running SAPT for each frame of a MD simulation is only possible with MD-SAPT toolkit;

- Psi4 has a wide range of options for running SAPT calculations (https://psicode.org/psi4manual/master/sapt.html#sapt0): for example, the default level of theory is SAPT0, but we also can run using SAPT2, SAPT2+ or SAPT2+3. Is it possible to change those options through MD-SAPT? If so, maybe the preformatted file should contain all possible settings with the standart values (since it is said on the paper that "This provides users with a preformatted file, such that they only need to fill in their parameters without any additional formatting").

- The example given in the paper is about the interactions between residuals in the Adenylate kinase. However, it seems that those kind of computations should be carried out using `fisapt` module (see https://psicode.org/psi4manual/master/fisapt.html#i-sapt-a-representative-example). It uses `sapt0` as underlying methodology, but extends its capability. Having this in mind: how MD-SAPT prepares the input for running Psi4? Is it already using `fisapt`? If not, how the user should change this?

- In the line 29 of page 1, it is said: "offering detailed insight into the strength and type of noncovalent interactions interactions". It should be discussed why this understanding is important and how it could be applied for a non-specialist audience. Maybe also cite references https://doi.org/10.1063/5.0127106 and https://doi.org/10.1021/acs.jcim.2c00788;

- In line "This result was generated as a demonstration of the kinds of data MD-SAPT can generate". This phrasing can lead the reader to think there are other types of data which can be generated by MD-SATP. If it is the case, the authors should explain all the outputs that can be acquired.

- Is there any plans to add support to other topology formats, e.g. amber topology formats (.prmtop)?
  - It seems one can use ParmEd to convert from .prmtop to .psf (https://github.com/ParmEd/ParmEd/issues/631)... I will try this as soon as I get MD-SAPT running at my end, but it would be nice to mention that the user can convert to get the needed file format.

## General comments

- The last commit to the MD-SAPT repository was made Oct 15, 2022. Is there any updates made after this?
