### 2023-11-03 095e9d2
Fix pydantic update problems in conda environment and config.py
