### 2023-11-05 7b7382b
Remove deprecated calls to parse_obj_as
