### 2023-11-08 calpolyccg/MDSAPT#43 @ALescoulie
@CarvFS, @ifd3f refactored our code to use a new version of pydantic, one issue was that a Psi4 depenency also used pydantic, but used a legacy version. A change I just merged should fix this issue. I'll make new release today as well
