### 2023-11-14 calpolyccg/MDSAPT#43 @CarvFS
Sorry about the delay! So, I have tried again and I am still getting errors because of `pydantic`:

```
/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/pydantic/_internal/_config.py:317: UserWarning: Valid config keys have changed in V2:
* 'allow_mutation' has been removed
  warnings.warn(message, UserWarning)
Traceback (most recent call last):
  File "/home/felipe/miniconda3/envs/mdsapt3/bin/mdsapt", line 5, in <module>
    from mdsapt.cli import cli
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/mdsapt/__init__.py", line 9, in <module>
    from .sapt import TrajectorySAPT, DockingSAPT
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/mdsapt/sapt.py", line 34, in <module>
    import psi4
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/psi4/__init__.py", line 76, in <module>
    from .driver import endorsed_plugins
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/psi4/driver/__init__.py", line 42, in <module>
    from qcelemental import constants
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/qcelemental/__init__.py", line 6, in <module>
    from . import covalent_radii, models, molparse, molutil, periodic_table, physical_constants, util, vanderwaals_radii
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/qcelemental/models/__init__.py", line 10, in <module>
    from .align import AlignmentMill
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/qcelemental/models/align.py", line 6, in <module>
    from ..util import blockwise_contract, blockwise_expand
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/qcelemental/util/__init__.py", line 1, in <module>
    from .autodocs import auto_gen_docs_on_demand, get_base_docs
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/qcelemental/util/autodocs.py", line 6, in <module>
    from pydantic import BaseModel, BaseSettings
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/pydantic/__init__.py", line 218, in __getattr__
    return _getattr_migration(attr_name)
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/pydantic/_migration.py", line 294, in wrapper
    raise PydanticImportError(
pydantic.errors.PydanticImportError: `BaseSettings` has been moved to the `pydantic-settings` package. See https://docs.pydantic.dev/2.4/migration/#basesettings-has-moved-to-pydantic-settings for more details.
```

As I understood, it seems something being used was moved in the second version of pydantic. I have tried using the first version of pydantic, but I got another error:

```
Traceback (most recent call last):
  File "/home/felipe/miniconda3/envs/mdsapt3/bin/mdsapt", line 5, in <module>
    from mdsapt.cli import cli
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/mdsapt/__init__.py", line 8, in <module>
    from .config import Config, load_from_yaml_file
  File "/home/felipe/miniconda3/envs/mdsapt3/lib/python3.9/site-packages/mdsapt/config.py", line 20, in <module>
    from pydantic import BaseModel, Field, model_validator, \
ImportError: cannot import name 'model_validator' from 'pydantic'
```

If I got it correctly, `model_validator` was added in the second version of pydantic.

Is there any specific version I should use?
