### 2023-11-20 calpolyccg/MDSAPT#43 @wiederm
Hello everyone!

I am here through https://github.com/openjournals/joss-reviews/issues/5931 :

I tried to install the package using
`mamba install -c psi4/label/dev -c conda-forge mdsapt`
with `python=3.11`, but the packages can't be resolved.

Also, could you please link the installation instructions or the documentation in the `README.md` or the sidebar?
