### 2023-11-20 #5931 @wiederm
# Review

Thank you for providing this software.

## Software

- Can you please re-enable the GH action tests to make sure that tests are passing?
- Unfortunately, I am having trouble installing the package with `python=3.11`. I added this to the issue: https://github.com/calpolyccg/MDSAPT/issues/43#issuecomment-1818803466
- I'd love to see more documentation. Unfortunately, documentation is currently limited to the example provided. Please provide more details on the various use cases of this analysis, the options to control the package, and the cli. Also, as far as I can tell, the documentation is only accessible by going through the `docs` badge. Please provide a link in the README.

## Manuscript

l40: comment needs to be removed
l53: comment needs to be resolved

- I'd love to hear more about the theory behind SAPT, why it is helpful to re-analyze MD simulations with different energy functions, and what kind of *scientific* questions can be answered with this toolkit. Please provide an example so that it is easier for the reader to understand the tool's purpose.
- Can you please describe (and document in the documentation) each option shown in Fig 1?
- Can you provide a guideline for choosing an appropriate QM level of theory?
