### 2023-11-20 calpolyccg/MDSAPT#43 @ifd3f
@CarvFS Did you try reinstalling the environment file with `mamba env update -f environment.yml`? I updated the dependencies. If that doesn't work, try deleting the environment and reinstalling it.

@wiederm The package currently in the psi4 repos is a year out of date. I'm working on publishing the new version right now.
