### 2023-11-20 calpolyccg/MDSAPT#43 @CarvFS
Hi @ifd3f! Thank you for your reply. I have tried your suggestiong, but I got

```
$ conda env update -f environment.yml
Retrieving notices: ...working... done
Solving environment: \ Killed
```
