### 2023-11-21 calpolyccg/MDSAPT#43 @ifd3f
Oh, right, the environment is extremely big because of how many dependencies both psi4 and mdsapt both have. Conda often uses lots of memory and runs forever to solve this environment, but mamba is much faster and more reliable.
