### 2023-11-21 calpolyccg/MDSAPT#43 @CarvFS
I have used `libmamba` solver (`conda env update -f environment.yml --solver=libmamba`) and it worked! Now the code is running, although with the warning:

```
Warning: importing 'simtk.openmm' is deprecated.  Import 'openmm' instead.
```

But MDSAPT is running now and the input file was generated as described in the submitted paper :)
