### 2023-11-21 #5931 @CarvFS
# Additional review:

## Generating input file:

- Generating the input file as described creates a file with the an empty `system_limits` field

```
system_limits:
  ncpus:
  memory:
```

and this cause an error. For example, by running:

```
mdsapt generate test.yaml
```

and then,

```
import mdsapt

config = mdsapt.load_from_yaml_file('test.yaml')
```

I have got the error:

```
2023-11-21 12:00:58,888 mdsapt.config ERROR    Error while loading config from 'test.yaml'
Traceback (most recent call last):
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/mdsapt/config.py", line 309, in load_from_yaml_file
    return Config.model_validate(yaml.safe_load(file))
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/pydantic/main.py", line 503, in model_validate
    return cls.__pydantic_validator__.validate_python(
pydantic_core._pydantic_core.ValidationError: 2 validation errors for Config
system_limits.ncpus
  Input should be a valid integer [type=int_type, input_value=None, input_type=NoneType]
    For further information visit https://errors.pydantic.dev/2.5/v/int_type
system_limits.memory
  Input should be a valid string [type=string_type, input_value=None, input_type=NoneType]
    For further information visit https://errors.pydantic.dev/2.5/v/string_type
Traceback (most recent call last):
  File "/mnt/d/OneDrive - Universidade Federal de Minas Gerais/Pós doc/paper-review/MD-SAPT/testing/run_mdsapt.py", line 3, in <module>
    config = mdsapt.load_from_yaml_file('test.yaml')
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/mdsapt/config.py", line 312, in load_from_yaml_file
    raise err
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/mdsapt/config.py", line 309, in load_from_yaml_file
    return Config.model_validate(yaml.safe_load(file))
  File "/home/felipe/miniconda3/envs/mdsapt/lib/python3.9/site-packages/pydantic/main.py", line 503, in model_validate
    return cls.__pydantic_validator__.validate_python(
pydantic_core._pydantic_core.ValidationError: 2 validation errors for Config
system_limits.ncpus
  Input should be a valid integer [type=int_type, input_value=None, input_type=NoneType]
    For further information visit https://errors.pydantic.dev/2.5/v/int_type
system_limits.memory
  Input should be a valid string [type=string_type, input_value=None, input_type=NoneType]
    For further information visit https://errors.pydantic.dev/2.5/v/string_type
```

By filling `ncpus` and `memory` it ran without any issue. Since those parameters will vary from one user to another, it could be filled with a general case (ncpus = 1 and memory = '1Gb', for example) or explicit tell the user to fill those.

The way it is described in the paper, the reader will be lead to think that one just need to run `mdsapt generate [filename]` and not add/modify anything else before running the calculations.

- The `generate` command must creates a file for the user. Thus, maybe it would be a good idea to add some required flags for `pairs`, `topology` and `trajectory`, since those are very specific options.

## Running SAPT calculations

- `sapt_run.run(config.start, config.stop, config.step)` throws an error:

```
 raise AttributeError(f'{type(self).__name__!r} object has no attribute {item!r}')
AttributeError: 'Config' object has no attribute 'start'
```

To make it run I have changed to

```
sapt_run.run(config.analysis.frames.start, config.analysis.frames.stop, config.analysis.frames.step)
```

This should be fixed in the paper example.
