### 2023-11-21 calpolyccg/MDSAPT#43 @ALescoulie
@CarvFS that's great! Conda has been such a headache this whole project since we have a large environment. Were also working on a conda-forge deployment that should be done this weekend. If you have any other question don't hesitate to ask @ifd3f and I are happy to help. You can follow issue #45 to see the progress on that.
