### 2024-01-18 #5931 @ifd3f
I'm currently working on packaging it with conda-forge (https://github.com/conda-forge/staged-recipes/pull/24782). We were previously packaged with psi4's channel but since they're distributing with conda-forge now, we're moving our package there too. I've been very busy with work recently, and had to deal with a cold and strep, but I will get back to packaging this soon.
