+++ build.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-# Used while building the conda package.
-python setup.py install
