### 2024-04-09 #5931 @ifd3f
It was finally merged into conda-forge after a lengthy process. I'll update the instructions in the paper once I get a chance.
