### 2024-05-04 0cc95ef
Update paper to reflect conda-forge channel
