### 2024-05-04 1518678
Document new install process using conda-forge
