### 2024-06-11 #5931 @ALescoulie
I wanted to apologize  for the delays, I've been dealing with some mental and physical health issues, but am committed to seeing this process through. I started my response to the reviews today, and will be putting out updates to the software and paper throughout the week. So It's easier to read, and I don't blow up everyone's notifications, I'll give the responses as a few larger comments on the thread.
