### 2024-06-11 fcca552
add more detailed descriptions to input file fields
