### 2024-06-11 08e682a
add more psi4 reference infomation
