### 2024-06-11 0a65fdc
add poetry file
