### 2024-06-12 67f902e
poetry2nix config with broken scipy
