+++ pyproject.toml
@@ -8,6 +8,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.9"
+cython = "^3.0.8"
 mdanalysis = "^2.7"
 click = "^8.1.7"
 rdkit = "^2023.9.6"
