### 2024-06-13 97b471e
get mdanalysis to build with nix
