### 2024-06-13 1b77806
update docs version and update toctree names
