+++ docs/index.rst
@@ -14,13 +14,12 @@ For the poster we presented at ACS Spring 2022, `click here <./_static/mdsapt_po
 
 
 .. toctree::
-   :maxdepth: 4
-   :numbered:
+   :maxdepth: 6
    :hidden:
 
    install
    quick
+   config
    sapt
-   optimizer
-   reader
-   viewer
+   repair
+
