+++ docs/viewer.rst
@@ -1,1 +1,0 @@
-.. automodule:: mdsapt.viewer
