+++ docs/api.rst
@@ -1,5 +1,0 @@
-API Documentation
-=================
-
-.. autosummary::
-   :toctree: autosummary
