### 2024-06-13 91bd0cb
fix strange toctree in docs
