+++ docs/repair.rst
@@ -1,0 +1,2 @@
+.. automodule:: mdsapt.repair
+
