### 2024-06-13 151b1d8
add install info to docs
