### 2024-06-13 #5931 @ALescoulie
# Response to First @CarvFS  Review

This is a summary of my response to the feedback in the first review in updates both to the paper and MD-SAPT its self.
I'm looking forward to hearing your feedback on my response, and working to improve this project further.

## Addressing Instigation issues

@ifd3f has worked on resolving the installation issues, and getting an official Conda-Forge release.
I also have updated the README and docs to reflect the following updates to installation.

> Python version 3.7 or older gives the error: "cannot import name 'Literal' from 'typing'";

We dropped 3.7 support, but I think we must have forgotten to remove that from out list of supported versions.
It was originals supported but in our 2.0.0 update, we used features from Python's Typing library that broken MD-SAPT for that release.

> Trying to install as conda install -c psi4/label/dev -c conda-forge mdsapt is not working

The Conda-Forge release should have resolved, you can install with MDSAPT using `conda install -c conda-forge mdsapt.
The root validator deprication issue was also fixed by @ifd3f.

## Addressing Important Links

> Link to readthedocs page at the paper is not redirecting to the correct page;

I fixed the links in the paper.

> The binder demo are returning some error messages.

I updated the demo's repo to use conda release of MD-SAPT, but for some reason it installs a borken version of psi4.
Later today I'll work on debugging that more, but for the time being its still broken.
The demo was first made for a confrence prior to MD-SAPT having a conda release so it was a bit hacked together, but now that we have a conda release it was easy to get up and running again.

## Addressing Paper and Toolkit Concerns

> When it says "including" in line 35 of page 1, it gives the ideia that there are other tasks not described.
> Is this the case? Maybe describe the all workflow tasks as bullet points in the paper and/or add it to the documentation;

I updated the information about the steps out workflow includes to be more clear. It now reads:

```
It combines the tasks required for this workflow:
  - selecting the specific residues for the QM calculation, adding protons where the residues were bound to the polypeptide chain
  - determining the formal charge
  - passing the charge and coordinates into psi4 (the QM calculation program)
  - saving the results into an easily queriable csv file or Pandas DataFrame
```

Do you think that it describes the process more clearly?

> It is said that a number of python based analysis tools exist for MD data.
> However, it is not clear if any of those also performs the same tasks implemented in MD-SAPT.
> For example: psi4 is used to run the quantum calculations, MDAnalysis is used to handle the MD simulation data.
> However, is not said running SAPT for each frame of a MD simulation is only possible with MD-SAPT toolkit;

I think that paragraph was unclear, we were indeding to discuss existing tools for working with MD data, and how MD-SAPT fits in with those.
None of them from my research support more involved analysis methods like what we implemented in MDSAPT.
In that section I was trying to convey what the MD ecosystem consists of, and how our project fits into that.
I rewrote that paragraph to make that more clear, and emphisize that we fit in as a pacakge built on top of an existing analysis system.
Is the following paragraph more clear?

```
Currently there are a number of libraries that simplify the process of studying MD data in software, such as MDtraj, pyLOOS, and MDAnalysis [@mcgibbon; @romo; @michaud; @gowers].
They provide a clean, common interface for reading, quering, and writing MD varrious file formats, that other analysis software can be built upon.
The Python library MDAnalysis is a core dependency of MD-SAPT.
By using MDAnalysis as a starting point for MD-SAPT, building our own system for reading MD data and select specified residues was not required.
```

The next paragraph elaborates on how MD-SAPT is an MDAKit, which is a part of the MDAnalysis ecosystem.

> Psi4 has a wide range of options for running SAPT calculations (https://psicode.org/psi4manual/master/sapt.html#sapt0): for example, the default level of theory is SAPT0, but we also can run using SAPT2, SAPT2+ or SAPT2+3.
> Is it possible to change those options through MD-SAPT?

Yes, it is by modifying the `psi4:method:` field of the input file.

> If so, maybe the preformatted file should contain all possible settings with the standart values (since it is said on the paper that "This provides users with a preformatted file, such that they only need to fill in their parameters without any additional formatting").

That is a good idea for making it more clear how to use MD-SAPT so I added some information to the input files generated using `mdsapt generate`.

This is the new format of the trajectory analysis file:

``` yaml
psi4:
  # The sapt theory level eg: SAPT0, SAPT2, SAPT2+, SAPT2+3
  # See https://psicode.org/psi4manual/master/sapt.html#sapt-level for more information
  method: 'sapt0'
  # Basis set for Psi4 SAPT calculation
  # See https://psicode.org/psi4manual/master/basissets_byelement.html#apdx-basiselement for complete list
  basis: 'jun-cc-pvdz'
  # set to true to save Psi4 output files
  save_output: true
  # Additional Psi4 settings see https://psicode.org for more psi4 specific documentation
  settings:
    # The reference wave function type eg: RHF, ROHF, UHF, CUHF, RKS, UKS
    # See https://psicode.org/psi4manual/master/autodir_options_c/scf__reference.html for more information
    reference: 'rhf'

simulation:
  ph: 7.0
  # pH of simulated system, it informs how protons are replaced when ballencing spin state before Psi4 calculations
  charge_guesser: 'standard'
  # charge_guesser: 'rdkit'  # to use rdkit. Make sure it is installed first.

system_limits:
  # Integer number of cpu cores to use in Psi4 calculation
  ncpus:
  # Given memory form `"{number_of_gigabytes}GB"`
  memory:

analysis:
  # This section is for running TrajectorySAPT. To run other types of analyses, see below.
  type: 'trajectory'

  topology: 'your/file/path.pdb'
  # If you want to override the charges of specific, you may specify it this way.
  # topology:
  #    path: 'your/file/path.pdb'
  #    charge_overrides:
  #      132: -2

  trajectories:
    - 'your/trajectory/path.dcd'
  pairs:
    # List the pairs of residues to calculate interaction energy between based on residue ids in topology file
    - [132, 152]
    - [34, 152]
  frames:
    # Starting frame
    start: 10

    # Stop frame
    stop: 30

    # frames per calculation step
    step: 1

  # Name of outputed csv file
  output: 'output.csv'

```

This is what the default docking input file gives you:

``` yaml
psi4:
  # The sapt theory level eg: SAPT0, SAPT2, SAPT2+, SAPT2+3
  # See https://psicode.org/psi4manual/master/sapt.html#sapt-level for more information
  method: 'sapt0'
  # Basis set for Psi4 SAPT calculation
  # See https://psicode.org/psi4manual/master/basissets_byelement.html#apdx-basiselement for complete list
  basis: 'jun-cc-pvdz'
  # set to true to save Psi4 output files
  save_output: true
  # Additional Psi4 settings see https://psicode.org for more psi4 specific documentation
  settings:
    # The reference wave function type eg: RHF, ROHF, UHF, CUHF, RKS, UKS
    # See https://psicode.org/psi4manual/master/autodir_options_c/scf__reference.html for more information
    reference: 'rhf'

simulation:
  ph: 7.0
  # pH of simulated system, it informs how protons are replaced when ballencing spin state before Psi4 calculations
  charge_guesser: 'standard'
  # charge_guesser: 'rdkit'  # to use rdkit. Make sure it is installed first.

system_limits:
  # Integer number of cpu cores to use in Psi4 calculation
  ncpus:
  # Given memory form `"{number_of_gigabytes}GB"`
  memory:

analysis:
  type: 'docking'
  # MD-SAPT analysis type (in this case docking)
  topology_directory: 'mek/docking'
  # The set of topology files for docking analysis
  pairs:
  # Place pair names defined above in list in a list
  - [15, 132]

system_limits:
  # Integer number of cpu cores to use in Psi4 calculation
  ncpus:
  # Given memory form `"{number_of_gigabytes}GB"`
  memory:

```
> The example given in the paper is about the interactions between residuals in the Adenylate kinase.
> However, it seems that those kind of computations should be carried out using fisapt module

I think this is a question that is could be better anwsered by @armcdona, as she is a quantum chemist and far more knowladgable on psi4 than I am.
I'll get her input on that question.

> Having this in mind: how MD-SAPT prepares the input for running Psi4?
> Is it already using fisapt?
> If not, how the user should change this?

I'll run some tests but looking at the Psi4 documentation, I see no reason why providing `method: fisapt` in an input file wouldn't work.
From looking at the docs it seems like this Psi4 input files are the same.

> In the line 29 of page 1, it is said: "offering detailed insight into the strength and type of noncovalent interactions interactions".
> It should be discussed why this understanding is important and how it could be applied for a non-specialist audience.

Thank you for the papers you attached with this feedback, they were very helpful for demonstrating the utility of MD-SAPT's methodology.
I added to the second paragraph of "Statment of Need" discussing the ACS-JCIM paper, and how they used SAPT calculations, to idenfity more residue interactions and identify interaction energies that differed sustantally from those predicted by MD.

Here is what I wrote verbatium:

"The molecular mechanics methods used in MD simulations do not characterize noncovalent interactions at the same level of detail as quantum mechanical methods, but allow simulations of large biomolecular systems [@lipkowitz].
In this work, we present a workflow tool to utilize ab initio methods as a post-processing technique on MD data, offering detailed insight into the strength and type of noncovalent interactions.
Using classical methods, such as those applied in MD, there are some noncovalent interactions that cannot be accurately modeled.
One paper that applied SAPT calculations to MD simulated proteins found that the SAPT calculations highlighted more residue interactions that shown in MD energy calculations, and that SAPT identified binding energies that differed from those predicted classically [@ding].
Their work illustrates the utility of SAPT in identifying critical active site interactions, and how that information can guide the search for ligands.
"

I hope that this better describes why MD-SAPT may be useful to an audience outside of quantum chemistry.

> In line "This result was generated as a demonstration of the kinds of data MD-SAPT can generate".
> This phrasing can lead the reader to think there are other types of data which can be generated by MD-SATP.
> If it is the case, the authors should explain all the outputs that can be acquired.

I argee that was vague wording, and I made it clear that the data in that figure is what MD-SAPT generates.

This is the replacement sentance:
"This result illistrates the non-covalent interaction energy decomposition data generated by MD-SAPT."

> Is there any plans to add support to other topology formats, e.g. amber topology formats (.prmtop)?

I have not specifically tested it but we can use any format that MDAnalysis supprots, so amber's .prmtop should work with current versions of MDAnalysis.
Here is their current format overview https://userguide.mdanalysis.org/stable/formats/index.html.

## Other Questions

> The last commit to the MD-SAPT repository was made Oct 15, 2022. Is there any updates made after this?

We have since then made some changes to the repository based on reviews and fixing releases.
