+++ mdsapt/_version.py
@@ -20,6 +20,8 @@ import functools
 
 # pylint: disable=consider-using-f-string
 # pylint: disable=invalid-name
+
+
 def get_keywords() -> Dict[str, str]:
     """Get the keywords needed to look up the version information."""
     # these strings will be replaced by git during git-archive.
