### 2024-06-21 ee0e6cc
try updating miniconda and removing psi4 channel
