### 2024-06-21 a756159
fix conda env creation error in CI pipeline
