### 2024-06-21 2856280
specify macos architecture
