### 2024-06-21 e7a6fe1
drop Python 3.8 from the CI
