+++ .github/workflows/CI.yaml
@@ -22,7 +22,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [macOS-latest, ubuntu-latest, macOS-13]
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     steps:
     - uses: actions/checkout@v4
@@ -84,7 +84,7 @@ jobs:
       fail-fast: true
       matrix:
         os: [macOS-13, ubuntu-latest]
-        python-version: [3.8, 3.9, 3.10, 3.11, 3.12]
+        python-version: [3.9, 3.10, 3.11, 3.12]
         include:
           - os: macOS-latest
             output-folder: osx-64
