### 2024-06-21 caf13fd
add python version to package and verify
