+++ .github/workflows/CI.yaml
@@ -139,8 +139,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [macOS-latest, ubuntu-latest]
-        python-version: [3.7, 3.8, 3.9]
+        os: [macOS-latest, macOS-13, ubuntu-latest]
+        python-version: [3.9, 3.10, 3.11, 3.12]
 
     steps:
     - uses: actions/checkout@v2
