### 2024-06-22 73e10a1
update pyproject to allow python 3.12
