+++ pyproject.toml
@@ -7,7 +7,7 @@ license = "GPL-3.0"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">3.9 <=3.12"
+python = "^3.9"
 #numpy = "1.26.1"
 mdanalysis = "^2.7"
 click = "^8.1.7"
