### 2024-06-22 4a289a1
fix CI and codecov button links to direct to CCG repo
