### 2024-06-24 6d131e3
fix cli message
