### 2025-02-23 #5931 @CarvFS
Hello everyone: I have started looking at the package again and noticed a coupled of things:

1 - In the quick start documentation (https://mdsapt.readthedocs.io/en/latest/quick.html), there is no changes regarding my additional comments on how to reference the start, stop and step variables: it works with `config.analysis.frames.start` and there is still showing `config.start`;

2 - The generated input file I have got from installing the package as `conda install -c conda-forge mdsapt` is:

```
psi4:
  method: 'sapt'
  basis: 'jun-cc-pvdz'
  save_output: true
  settings:
    reference: 'rhf'
simulation:
  ph: 7.0
  charge_guesser: 'standard'
  # charge_guesser: 'rdkit'  # to use rdkit. Make sure it is installed first.
system_limits:
  ncpus:
  memory:
analysis:
  ### This section is for running TrajectorySAPT. To run other types of analyses, see below.
  type: 'trajectory'

  topology: 'your/file/path.pdb'
  # If you want to override the charges of specific, you may specify it this way.
  # topology:
  #    path: 'your/file/path.pdb'
  #    charge_overrides:
  #      132: -2

  trajectories:
    - 'your/trajectory/path.dcd'
  pairs:
    - [132, 152]
    - [34, 152]
  frames:
    start: 10
    stop: 30
    #step: 2

  output: 'output.csv'

#  ### For running DockingSAPT
#  type: 'docking'
#  mode: 'protein-ligand'
#  topologies:
#    - 'one/topology.pdb'
#    - 'another/topology.pdb'
#    - 'a/glob/of/topologies/*.pdb'
#    - 'a/directory/full/of/topologies'
#    - path: 'one/specific/topology/with/weird/charges.pdb'
#      charge_overrides:
#        145: -4
#
#  pairs:
#    - [132, 152]
#    - [34, 152]
#
#  output: 'output.csv'

```

The template files in the repository are updated. However, through conda install the user would not get the most recent version. I could install from the source, but I believe the most users will want to install using conda. So, I think I will wait until the authors fix this to update my review.
