### 2025-04-03 #5931 @arfon
@armcdona @ALescoulie @ifd3f – I think we're waiting on your input again here regarding some conda issues again. Please take a look and see what you can do to unblock this.
