### 2025-04-25 #5931 @ifd3f
Hi, sorry for the really late response. I've been busy with work and other stuff in my life, so I haven't been keeping up with this. It seems like the conda package broke again in between the review cycles. I'll let you know when I have it fixed.
