### 2025-04-26 261b095
fix rc suffix not accepted by conda
