### 2025-07-14 #5931 @CarvFS
Sorry about the delay!

# Updated review

The example in the paper and quick reference are still showing

```
sapt_run.run(config.start, config.stop, config.step)
```

but this is not working. I have to run as I have mentioned previously:

```
sapt_run.run(config.analysis.frames.start, config.analysis.frames.stop, config.analysis.frames.step)
```

> note that `guess_atom_element` is deprecated and will be removed in release 3.0.0. So, just be cautions on future updates :)

So, conda install and running the calculations (with the changes mentioned above) seems to be fine. Previously I have suggested changes to the manuscript as well, but I am seeing the same version as before while downloading the manuscript here. Was it not updated or am I missing something?

Thanks!
