### 2025-09-18 #5931 @armcdona
We are not going to continue to pursue publication with JOSS.  Please withdraw our submission.
